changes: update and delete contact

+++ controllers/contactController.js
@@ -58,6 +58,12 @@ module.exports.updateContact = asyncHandler(async (req, res) => {
 		res.status(404);
 		throw new Error('Contact not found');
 	}
+
+	if (contact.user_id.toString() !== req.user.id) {
+		res.status(403);
+		throw new Error('User deos not have permission to delete this contact');
+	}
+
 	const updatedContact = await Contact.findByIdAndUpdate(
 		req.params.id,
 		req.body,
@@ -74,6 +80,10 @@ module.exports.deleteContact = asyncHandler(async (req, res) => {
 	if (!contact) {
 		res.status(404);
 		throw new Error('Contact not found');
+	}
+	if (contact.user_id.toString() !== req.user.id) {
+		res.status(403);
+		throw new Error('User deos not have permission to delete this contact');
 	}
 	await Contact.findByIdAndDelete(req.params.id);
 	res.status(201).send(contact);
